+++ src/AddOperation.java
@@ -1,6 +1,6 @@
 public class AddOperation extends AbstractOperation{
     @Override
-    public double operate(int firstNumber, int secondNumber){
+    public double operate(float firstNumber, float secondNumber){
         return firstNumber + secondNumber;
     }
 }
